refactor(desktop): split start session shared helpers
