+++ apps/desktop/src/state/operations/agent-orchestrator/handlers/start-session-constants.ts
@@ -1,0 +1,4 @@
+export const STALE_START_ERROR = "Workspace changed while starting session.";
+
+export const MISSING_BUILD_TARGET_ERROR =
+  "Builder continuation cannot start until a builder worktree exists";
