+++ apps/desktop/src/state/operations/agent-orchestrator/handlers/start-session-fork-strategy.ts
@@ -6,14 +6,10 @@ import type {
   StartSessionCreationInput,
   StartSessionExecutionDependencies,
 } from "./start-session.types";
+import { registerStartedSession, stopSessionOnStaleAndThrow } from "./start-session-persistence";
+import { assertScenarioStartPolicy, resolveStartTask } from "./start-session-policies";
 import { resolveLoadedSourceSession } from "./start-session-reuse-strategy";
-import {
-  assertScenarioStartPolicy,
-  registerStartedSession,
-  resolveRuntimeAndModel,
-  resolveStartTask,
-  stopSessionOnStaleAndThrow,
-} from "./start-session-shared";
+import { resolveRuntimeAndModel } from "./start-session-runtime";
 
 type ForkStrategyInput = {
   ctx: StartSessionContext;
