+++ apps/desktop/src/state/operations/agent-orchestrator/handlers/start-session-runtime.ts
@@ -1,0 +1,131 @@
+import type { TaskCard } from "@openducktor/contracts";
+import type { AgentModelSelection, AgentScenario } from "@openducktor/core";
+import { errorMessage } from "@/lib/errors";
+import { normalizeWorkingDirectory, throwIfRepoStale } from "../support/core";
+import { inferScenario } from "../support/scenario";
+import { createSessionPromptContext, loadSessionPromptInputs } from "../support/session-prompt";
+import type {
+  ResolvedRuntimeAndModel,
+  RuntimeDependencies,
+  StartSessionContext,
+  StartSessionExecutionDependencies,
+} from "./start-session.types";
+import { MISSING_BUILD_TARGET_ERROR, STALE_START_ERROR } from "./start-session-constants";
+
+export const resolveRuntimeAndModel = async ({
+  ctx,
+  scenario,
+  targetWorkingDirectory,
+  requestedRuntimeKind,
+  taskCard,
+  deps,
+}: {
+  ctx: StartSessionContext;
+  scenario: AgentScenario | undefined;
+  targetWorkingDirectory?: string | null;
+  requestedRuntimeKind?: AgentModelSelection["runtimeKind"] | null;
+  taskCard: TaskCard;
+  deps: Pick<StartSessionExecutionDependencies, "runtime" | "task" | "model">;
+}): Promise<ResolvedRuntimeAndModel> => {
+  const { documents: docs, promptOverrides } = await loadSessionPromptInputs({
+    repoPath: ctx.repoPath,
+    taskId: ctx.taskId,
+    loadTaskDocuments: deps.task.loadTaskDocuments,
+    loadRepoPromptOverrides: deps.model.loadRepoPromptOverrides,
+  });
+  throwIfRepoStale(ctx.isStaleRepoOperation, STALE_START_ERROR);
+
+  const resolvedScenario = scenario ?? inferScenario(ctx.role, taskCard, docs);
+  throwIfRepoStale(ctx.isStaleRepoOperation, STALE_START_ERROR);
+
+  const runtimeInfo = await deps.runtime.ensureRuntime(ctx.repoPath, ctx.taskId, ctx.role, {
+    ...(targetWorkingDirectory !== undefined ? { targetWorkingDirectory } : {}),
+    ...(requestedRuntimeKind ? { runtimeKind: requestedRuntimeKind } : {}),
+  });
+  throwIfRepoStale(ctx.isStaleRepoOperation, STALE_START_ERROR);
+
+  const { systemPrompt } = createSessionPromptContext({
+    role: ctx.role,
+    scenario: resolvedScenario,
+    task: taskCard,
+    promptOverrides,
+    documents: docs,
+  });
+
+  return {
+    taskCard,
+    runtime: runtimeInfo,
+    resolvedScenario,
+    systemPrompt,
+    promptOverrides,
+  };
+};
+
+export const resolveFreshStartTargetWorkingDirectory = async ({
+  ctx,
+  runtime,
+}: {
+  ctx: StartSessionContext;
+  runtime: RuntimeDependencies;
+}): Promise<string | null | undefined> => {
+  if (ctx.role === "qa") {
+    return runtime.resolveBuildContinuationTarget(ctx.repoPath, ctx.taskId);
+  }
+
+  if (ctx.role !== "build") {
+    return undefined;
+  }
+
+  try {
+    return await runtime.resolveBuildContinuationTarget(ctx.repoPath, ctx.taskId);
+  } catch (error) {
+    if (errorMessage(error).includes(MISSING_BUILD_TARGET_ERROR)) {
+      return null;
+    }
+    throw error;
+  }
+};
+
+export const resolveFreshStartTargetWorkingDirectoryForStart = async ({
+  ctx,
+  runtime,
+  targetWorkingDirectory,
+}: {
+  ctx: StartSessionContext;
+  runtime: RuntimeDependencies;
+  targetWorkingDirectory?: string | null;
+}): Promise<{
+  targetWorkingDirectory: string | null | undefined;
+  normalizedTargetWorkingDirectory: string;
+}> => {
+  if (targetWorkingDirectory !== undefined) {
+    return {
+      targetWorkingDirectory,
+      normalizedTargetWorkingDirectory: normalizeWorkingDirectory(targetWorkingDirectory),
+    };
+  }
+
+  const targetWorkingDirectoryForStart = await resolveFreshStartTargetWorkingDirectory({
+    ctx,
+    runtime,
+  });
+  return {
+    targetWorkingDirectory: targetWorkingDirectoryForStart,
+    normalizedTargetWorkingDirectory: normalizeWorkingDirectory(targetWorkingDirectoryForStart),
+  };
+};
+
+export const serializeSelectedModelKey = (
+  selectedModel: AgentModelSelection | undefined,
+): string => {
+  if (!selectedModel) {
+    return "";
+  }
+  return [
+    selectedModel.runtimeKind ?? "",
+    selectedModel.providerId,
+    selectedModel.modelId,
+    selectedModel.variant ?? "",
+    selectedModel.profileId ?? "",
+  ].join("::");
+};
